3.3.4 add unity terminal
