+++ .env
@@ -3,3 +3,4 @@ TM_LANGUAGE=https://unpkg.com/greyscript-textmate@1.8.0/dist/greyscript.tmLangua
 TM_LANGUAGE_CONFIG=https://unpkg.com/greyscript-textmate@1.8.0/dist/greyscriptLanguageConfig.json
 TM_THEME=https://unpkg.com/greyscript-textmate@1.8.0/dist/greyscript.theme.json
 EDITOR_SERVICE_URL=https://editor-dahpv.ondigitalocean.app
+GREYBEL_TERMINAL_URL=https://unpkg.com/greybel-terminal@0.1.1
