+++ web/index.css
@@ -232,7 +232,6 @@ body {
 	height: 200px;
 	background: #000;
 	color: green;
-	padding: 5px;
 	overflow: auto;
 	border: 1px solid #FFF;
 	position: relative;
