+++ web/index.html
@@ -10,6 +10,8 @@
       <link rel="icon" type="image/x-icon" href="./favicon.ico">
 			<link rel="canonical" href="https://editor.greyscript.org">
 			<link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+      <script defer crossorigin src="https://unpkg.com/greybel-terminal@0.1.1/dist/view.js"></script>
+      <script defer crossorigin src="https://unpkg.com/greybel-terminal@0.1.1/assets/preview.loader.js"></script>
 			<script defer crossorigin src="https://unpkg.com/react@18.2.0/umd/react.production.min.js"></script>
       <script defer crossorigin src="https://unpkg.com/react-dom@18.2.0/umd/react-dom.production.min.js"></script>
 	</head>
